+++ CHANGELOG.md
@@ -6,10 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Unreleased
 ************
-* [BUGFIX] fixing validation for the observing constraints
 * Adds a new POST /{prj_id}/{ob_id}/generateGSMSurveySBDefinitions endpoint that will generate SBDefinitions
   from the pointings file (currently stored in this repo as well) and other inputs.
 * Adding quality_attribute_metrics to low configurations from OSD
+* [BUGFIX] fixing constraint validation that did not correctly take the scan duration into account
+* [BUGFIX] fixing constraint validation that did not correctly account for LST constraints that span midnight
+
 
 14.3.0
 ************